more refactoring of proxy topics
